Fix older torch versions dtype casting bug

+++ preprocess/squeeze_by_class.py
@@ -24,7 +24,7 @@ def squeeze_by_classes(label, is_one_hot) -> torch.Tensor:
 
     else:
         for cls in all_classes:
-            mask = torch.where((label == cls) & (label > 0), torch.tensor(cls), torch.tensor(0)).squeeze()
+            mask = torch.where((label == cls) & (label > 0), torch.tensor(cls), torch.tensor(0, dtype=cls.dtype)).squeeze()
             masks.append(mask)
     one_hot = torch.stack(masks, dim=0)
     return one_hot
